Split up forms.test.tsx.
